+++ src-plugin/Plugin/Plugin.cs
@@ -18,9 +18,6 @@
         public required PluginConfig Config { get; set; } = new PluginConfig();
         public GameConfig? GameConfig { get; set; }
         public Menu.KitsuneMenu Menu { get; private set; } = null!;
-        public static readonly Random rng = new();
-        public static MemoryFunctionVoid<IntPtr, string, IntPtr, IntPtr, IntPtr, IntPtr, IntPtr, IntPtr>? GiveNamedItem2;
-
         public bool IsBetweenRounds = false;
         public bool HasDatabase = false;
 
@@ -32,14 +29,6 @@
             {
                 base.Logger.LogWarning("Configuration version mismatch (Expected: {0} | Current: {1})", this.Config.Version, config.Version);
             }
-
-            //** ? Signature Check */
-
-            GiveNamedItem2 = RuntimeInformation.IsOSPlatform(OSPlatform.Linux)
-                ? new(@"\x55\x48\x89\xE5\x41\x57\x41\x56\x41\x55\x41\x54\x53\x48\x83\xEC\x18\x48\x89\x7D\xC8\x48\x85\xF6\x74")
-                : RuntimeInformation.IsOSPlatform(OSPlatform.Windows)
-                    ? new(@"\x48\x83\xEC\x38\x48\xC7\x44\x24\x28\x00\x00\x00\x00\x45\x33\xC9\x45\x33\xC0\xC6\x44\x24\x20\x00\xE8\x2A\x2A\x2A\x2A\x48\x85")
-                    : null;
 
             //** ? Load Round Types */
 
@@ -62,7 +51,6 @@
 
         public CCSGameRules? gameRules = null;
         public Timer? WarmupTimer { get; set; } = null;
-        public Timer? ForceClanTimer { get; set; } = null;
 
         public override void Load(bool hotReload)
         {
@@ -117,7 +105,7 @@
 
             if (Config.CompatibilitySettings.ForceArenaClantags)
             {
-                ForceClanTimer = AddTimer(1, () =>
+                AddTimer(1, () =>
                 {
                     if (Arenas is null) return;
 
@@ -143,17 +131,6 @@
                     }
                 }, TimerFlags.REPEAT);
             }
-        }
-
-        public override void Unload(bool hotReload)
-        {
-            List<ArenaPlayer> players = Utilities.GetPlayers()
-                 .Where(p => p.IsValid && p.PlayerPawn?.IsValid == true && !p.IsBot && !p.IsHLTV && p.Connected == PlayerConnectedState.PlayerConnected)
-                 .Select(p => Arenas?.FindPlayer(p)!)
-                 .Where(p => p != null)
-                 .ToList();
-
-            Task.Run(() => SavePlayerPreferencesAsync(players));
         }
     }
 }