+++ src-plugin/Plugin/PluginCommands.cs
@@ -2,6 +2,7 @@ namespace K4Arenas
 {
 	using CounterStrikeSharp.API;
 	using CounterStrikeSharp.API.Core;
+	using CounterStrikeSharp.API.Core.Translations;
 	using CounterStrikeSharp.API.Modules.Commands;
 	using CounterStrikeSharp.API.Modules.Commands.Targeting;
 	using CounterStrikeSharp.API.Modules.Utils;
@@ -62,7 +63,7 @@ namespace K4Arenas
 
 			if (p1.Challenge is null)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.notchallenged"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.notchallenged")}");
 				return;
 			}
 
@@ -70,7 +71,7 @@ namespace K4Arenas
 
 			if (!p2.IsValid)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.notavailable"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.notavailable")}");
 				p1.Challenge = null;
 				return;
 			}
@@ -78,8 +79,8 @@ namespace K4Arenas
 			p1.Challenge!.IsAccepted = true;
 			p2.Challenge!.IsAccepted = true;
 
-			info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.accepted", p2.Controller.PlayerName]}");
-			p2.Controller.PrintToChat($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.acceptedby", player!.PlayerName]}");
+			info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.accepted", p2.Controller.PlayerName)}");
+			p2.Controller.PrintToChat($" {Localizer.ForPlayer(p2.Controller, "k4.general.prefix")} {Localizer.ForPlayer(p2.Controller, "k4.general.challenge.acceptedby", player!.PlayerName)}");
 		}
 
 		public void Command_Decline(CCSPlayerController? player, CommandInfo info)
@@ -94,7 +95,7 @@ namespace K4Arenas
 
 			if (p1.Challenge is null)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.notchallenged"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.notchallenged")}");
 				return;
 			}
 
@@ -102,7 +103,7 @@ namespace K4Arenas
 
 			if (!p2.IsValid)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.notavailable"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.notavailable")}");
 				p1.Challenge = null;
 				return;
 			}
@@ -110,8 +111,8 @@ namespace K4Arenas
 			p1.Challenge = null;
 			p2.Challenge = null;
 
-			info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.declined", p2.Controller.PlayerName]}");
-			p2.Controller.PrintToChat($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.declinedby", player!.PlayerName]}");
+			info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.declined", p2.Controller.PlayerName)}");
+			p2.Controller.PrintToChat($" {Localizer.ForPlayer(p2.Controller, "k4.general.prefix")} {Localizer.ForPlayer(p2.Controller, "k4.general.challenge.declinedby", player!.PlayerName)}");
 		}
 
 		public void Command_Challenge(CCSPlayerController? player, CommandInfo info)
@@ -126,14 +127,14 @@ namespace K4Arenas
 
 			if (p1.Challenge != null)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.inchallenge"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.inchallenge")}");
 				return;
 			}
 
 			TargetResult targetResult = info.GetArgTargetResult(1);
 			if (targetResult.Count() != 1)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.invalidtarget"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.invalidtarget")}");
 				return;
 			}
 
@@ -142,13 +143,13 @@ namespace K4Arenas
 
 			if (p2 is null || p2 == p1)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.invalidtarget"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.invalidtarget")}");
 				return;
 			}
 
 			if (p2.Challenge != null)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.inchallenge"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.inchallenge")}");
 				return;
 			}
 
@@ -157,7 +158,7 @@ namespace K4Arenas
 
 			if (p1ArenaID == -1 || p2ArenaID == -1)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.notinarena"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.notinarena")}");
 				return;
 			}
 
@@ -165,8 +166,8 @@ namespace K4Arenas
 			p1.Challenge = challenge;
 			p2.Challenge = challenge;
 
-			info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.waiting", challengedPlayer.PlayerName]}");
-			challengedPlayer.PrintToChat($" {Localizer["k4.general.prefix"]} {Localizer["k4.general.challenge.request", player!.PlayerName, Config.CommandSettings.ChallengeAcceptCommands.FirstOrDefault("Missing"), Config.CommandSettings.ChallengeDeclineCommands.FirstOrDefault("Missing")]}");
+			info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.general.challenge.waiting", challengedPlayer.PlayerName)}");
+			challengedPlayer.PrintToChat($" {Localizer.ForPlayer(challengedPlayer, "k4.general.prefix")} {Localizer.ForPlayer(challengedPlayer, "k4.general.challenge.request", player!.PlayerName, Config.CommandSettings.ChallengeAcceptCommands.FirstOrDefault("Missing"), Config.CommandSettings.ChallengeDeclineCommands.FirstOrDefault("Missing"))}");
 		}
 
 		public void Command_AFK(CCSPlayerController? player, CommandInfo info)
@@ -203,7 +204,7 @@ namespace K4Arenas
 				}
 			}
 
-			info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {(arenaPlayer.AFK ? string.Format(Localizer["k4.chat.afk_enabled"], Config.CommandSettings.AFKCommands.FirstOrDefault("Missing")) : Localizer["k4.chat.afk_disabled"])}");
+			info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {(arenaPlayer.AFK ? string.Format(Localizer.ForPlayer(player, "k4.chat.afk_enabled"), Config.CommandSettings.AFKCommands.FirstOrDefault("Missing")) : Localizer.ForPlayer(player, "k4.chat.afk_disabled"))}");
 		}
 
 		public void Command_Queue(CCSPlayerController? player, CommandInfo info)
@@ -215,11 +216,11 @@ namespace K4Arenas
 
 			if (queuePlace == -1)
 			{
-				info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.chat.queue_not_in_queue"]}");
+				info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.chat.queue_not_in_queue")}");
 				return;
 			}
 
-			info.ReplyToCommand($" {Localizer["k4.general.prefix"]} {Localizer["k4.chat.queue_position", queuePlace + 1]}");
+			info.ReplyToCommand($" {Localizer.ForPlayer(player, "k4.general.prefix")} {Localizer.ForPlayer(player, "k4.chat.queue_position", queuePlace + 1)}");
 		}
 
 		public void Command_RoundPref(CCSPlayerController? player, CommandInfo info)